+++ lib/models/participant.rb
@@ -1,23 +1,17 @@
 require 'date'
 
 class Participant
-  attr_reader :address
+  attr_reader :identifier
   attr_reader :balance
-  attr_reader :recent_winner
-  attr_reader :nft_rare_holder
-  attr_reader :nft_common_holder
 
-  def initialize(address:, balance:, recent_winner:, nft_rare_holder: false, nft_common_holder: false)
-    @address = address.downcase
-    @balance = (balance && balance >= 0) ? balance : 0 # guarding against invalid balances
-    @recent_winner = recent_winner
-    @nft_rare_holder = nft_rare_holder
-    @nft_common_holder = nft_common_holder
+  def initialize(identifier:, balance:, xorshift: nil)
+    @identifier = identifier
+    @balance    = (balance && balance >= 0) ? balance : 0 # guarding against invalid balances
+    @xorshift   = xorshift || XorshiftService.new
   end
 
-  TICKET_PRICE = 250.freeze                   # e.g: 100 means 1 ticket = 100 POLS
-  NO_COOLDOWN_MINIMUM_BALANCE = 30_000.freeze # minimum balance to avoid cooldown
-  BALANCE_WEIGHTS = {                         #  e.g: { 1000 => 1.1 } means 1000 POLS weigths 1.1
+  TICKET_PRICE = 250.freeze # e.g: 100 means 1 ticket = 100 POLS
+  BALANCE_WEIGHTS = {       # e.g: { 1000 => 1.1 } means 1000 POLS weigths 1.1
     0      => 0.00,
     250    => 1.00,
     1_000  => 1.10,
@@ -25,6 +19,14 @@ class Participant
     10_000 => 1.20,
     30_000 => 1.25
   }.freeze
+
+  def dice
+    @dice ||= @xorshift.next_float
+  end
+
+  def probability
+    dice ** (1.0 / tickets) # reservoir sampling
+  end
 
   def tickets
     (balance / TICKET_PRICE).to_i * weight
@@ -39,18 +41,10 @@ class Participant
   end
 
   def eligible?
-    return true if nft_rare_holder
-
-    tickets > 0 && !weight.nil? && !in_cooldown_period?
-  end
-
-  def in_cooldown_period?
-    return false if nft_rare_holder || nft_common_holder || balance >= NO_COOLDOWN_MINIMUM_BALANCE
-
-    recent_winner
+    tickets > 0
   end
 
   def <=>(other)
-    other.balance <=> self.balance # more balance comes first
+    other.probability <=> self.probability
   end
 end